FOIA-280: Move modal to filter component.
Moves the modal stuff to the typeahead's wrapper filter component since
the typeahead component already does a lot.
