+++ js/components/report_agency_component_typeahead.jsx
@@ -173,68 +173,22 @@ class ReportAgencyComponentTypeahead extends Component {
     });
   }
 
-  buildModalContent() {
-    const checkboxes = agencyComponentStore
-      .getAgencyComponentsForAgency(this.props.selectedAgency.id)
-      .map((component) => {
-        const inputId = uniqueId(`${component.abbreviation}_`);
-        return (
-          <li className="usa-width-one-third" key={component.id}>
-            <input
-              id={inputId}
-              type="checkbox"
-              name={`${this.props.selectedAgency.id}-component`}
-              value={component.id}
-            />
-            <label htmlFor={inputId}>{component.abbreviation}</label>
-          </li>
-        );
-      });
-    return (
-      <fieldset className="usa-fieldset-inputs">
-        <legend className="usa-sr-only">Select Agency Components</legend>
-        <ul className="usa-unstyled-list usa-grid checkbox-list checkbox-list--in-modal">
-          {checkboxes}
-        </ul>
-        <div className="form-group">
-          <ul className="inline-list">
-            <li><a href="#">Select All</a></li>
-            <li><a href="#">Select None</a></li>
-          </ul>
-        </div>
-      </fieldset>
-    );
-  }
-
   render() {
     const loading = !this.props.agencyFinderDataComplete;
-    const agencyIsSelected = (this.props.selectedAgency.id !== 0 && this.props.selectedAgency.type === 'agency') || false;
-    const isCentralizedAgency = this.props.selectedAgency.component_count <= 1 || false;
 
     return (
-      <div className="typeahead-search">
-        <div role="search">
-          <label className="usa-sr-only" htmlFor={`agency-component-search-${this.state.id}`}>Search for an agency</label>
-          <input
-            type="text"
-            id={`agency-component-search-${this.state.id}`}
-            name="agency_component_search"
-            placeholder={loading ? `Loading… ${this.props.agencyFinderDataProgress}%` : 'Type agency name'}
-            disabled={loading}
-            ref={(input) => {
-              this.typeaheadInput = input;
-            }}
-          />
-        </div>
-        {agencyIsSelected && !isCentralizedAgency &&
-          <div className="form-group">
-            <FoiaModal
-              modalContent={this.buildModalContent()}
-              ariaLabel="Filter agency components"
-              triggerText="Select Agency Components"
-            />
-          </div>
-        }
+      <div role="search">
+        <label className="usa-sr-only" htmlFor={`agency-component-search-${this.state.id}`}>Search for an agency</label>
+        <input
+          type="text"
+          id={`agency-component-search-${this.state.id}`}
+          name="agency_component_search"
+          placeholder={loading ? `Loading… ${this.props.agencyFinderDataProgress}%` : 'Type agency name'}
+          disabled={loading}
+          ref={(input) => {
+            this.typeaheadInput = input;
+          }}
+        />
       </div>
     );
   }
